Infrastructure for hiding backtrace frames
This implements a hide_in_backtrace meta which may be attached to
individual methods (via a flag on the CodeInfo) to mark them for hiding
in backtraces.

Also implement a heuristic function `StackTraces.frame_importance()`
which will take this information along with the originating module and
derive an "importance" for the frame. Use this importance level for
filtering, rather than directly using the from_c flag.